+++ core/src/main/java/com/ray3k/particleparkpro/PresetActions.java
@@ -13,7 +13,7 @@ import static com.badlogic.gdx.scenes.scene2d.actions.Actions.*;
 import static com.ray3k.particleparkpro.Core.*;
 
 /**
- * A convenience class to organize the Actions used for transitions throughout Particle Park Pro.
+ * A convenience class to organize the Actions used for transitions throughout GDX Particle Editor.
  */
 public class PresetActions {
     private final static Color CLEAR_WHITE = new Color(1, 1, 1, 0);