+++ core/src/main/java/com/ray3k/particleparkpro/widgets/poptables/PopImageError.java
@@ -43,7 +43,7 @@ public class PopImageError extends PopTable {
         pad(20);
         defaults().space(10);
 
-        var label = new Label("Particle Park Pro encountered an error:", skin, "bold");
+        var label = new Label("GDX Particle Editor encountered an error:", skin, "bold");
         add(label);
 
         row();