+++ core/src/main/java/com/ray3k/particleparkpro/widgets/tables/ClassicTable.java
@@ -55,7 +55,7 @@ public class ClassicTable extends Table {
         table.add(textButton).spaceLeft(10);
         addHandListener(textButton);
         addTooltip(textButton, "Open browser to download page", Align.top, Align.top, tooltipBottomArrowStyle);
-        onChange(textButton, () -> Gdx.net.openURI("https://github.com/raeleus/Particle-Park-Pro/releases"));
+        onChange(textButton, () -> Gdx.net.openURI("https://github.com/libgdx/gdx-particle-editor/releases"));
         Utils.checkVersion((String newVersion) -> {
             if (!versionRaw.equals(newVersion)) textButton.setVisible(true);
         });