+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 ### Version 0.0.5
-* 
+* Rebranded app as GDX Particle Editor
 ### Version 0.0.4
 * Keyboard shortcuts to save, saveAs, and open have been added. Thanks, John!
 * Added option to load pre-built templates